+++ src/modules/sbstudio/plugin/operators/update_time_markers_from_storyboard.py
@@ -1,3 +1,8 @@
+from json import dumps, loads
+from typing import List, Optional
+
+from sbstudio.plugin.utils.collections import filter_collection
+
 from .base import StoryboardOperator
 
 __all__ = ("UpdateTimeMarkersFromStoryboardOperator",)
@@ -11,14 +16,44 @@ class UpdateTimeMarkersFromStoryboardOperator(StoryboardOperator):
     bl_description = "Update all time markers to be synchronized with the storyboard"
 
     def execute_on_storyboard(self, storyboard, context):
-        markers = context.scene.timeline_markers
-        markers.clear()
+        our_marker_names: Optional[List[str]] = None
+
+        scene = context.scene
+
+        markers = scene.timeline_markers
+        our_marker_names_as_string = scene.skybrush.settings.time_markers
+        if our_marker_names_as_string:
+            try:
+                our_marker_names = loads(our_marker_names_as_string)
+            except Exception:
+                # it's okay, our_marker_names remains at None
+                pass
+
+        if our_marker_names is not None:
+            # Remove only those markers that we have created
+            is_not_our_marker = lambda marker: marker.name not in our_marker_names
+            filter_collection(markers, is_not_our_marker)
+        else:
+            markers.clear()
+
+        our_marker_names = []
         for entry in storyboard.entries:
-            markers.new("at {}".format(entry.name), frame=entry.frame_start)
+            marker_name = "at {}".format(entry.name)
+            our_marker_names.append(marker_name)
+            markers.new(marker_name, frame=entry.frame_start)
+
             if entry.duration > 0:
+                marker_name = "{} ends".format(entry.name)
+                our_marker_names.append(marker_name)
                 markers.new(
-                    "{} ends".format(entry.name),
+                    marker_name,
                     frame=entry.frame_start + entry.duration,
                 )
+
+        try:
+            scene.skybrush.settings.time_markers = dumps(our_marker_names)
+        except Exception:
+            # weird, but let's not fail
+            pass
 
         return {"FINISHED"}