+++ src/modules/sbstudio/plugin/model/settings.py
@@ -1,6 +1,15 @@
-from bpy.props import BoolProperty, EnumProperty, FloatProperty, PointerProperty
+from bpy.props import (
+    BoolProperty,
+    EnumProperty,
+    FloatProperty,
+    IntProperty,
+    PointerProperty,
+    StringProperty,
+)
 from bpy.types import Collection, PropertyGroup
 
+from sbstudio.math.rng import RandomSequence
+from sbstudio.plugin.constants import RANDOM_SEED_MAX
 from sbstudio.plugin.utils.bloom import set_bloom_effect_enabled
 
 
@@ -32,6 +41,15 @@ class DroneShowAddonFileSpecificSettings(PropertyGroup):
         soft_max=20,
     )
 
+    random_seed = IntProperty(
+        name="Random seed",
+        description="Root random seed value used to generate randomized stuff in this show file",
+        default=0,
+        min=1,
+        soft_min=1,
+        soft_max=RANDOM_SEED_MAX,
+    )
+
     show_type = EnumProperty(
         name="Show type",
         description="Specifies whether the drone show is an outdoor or an indoor show",
@@ -58,3 +76,29 @@ class DroneShowAddonFileSpecificSettings(PropertyGroup):
         default=True,
         update=use_bloom_effect_updated,
     )
+
+    time_markers = StringProperty(
+        name="Time markers",
+        description=(
+            "Names of the timeline markers that were created by the plugin and "
+            "that may be removed when the 'Update Time Markers' operation "
+            "is triggered"
+        ),
+        default="",
+        options={"HIDDEN"},
+    )
+
+    @property
+    def random_sequence_root(self) -> RandomSequence:
+        """Returns a random sequence generated from the random seed associated
+        to the project.
+
+        Do not hold on to a reference to the random sequence returned from this
+        property permanently; when the user changes the random seed, the old
+        instance will not be invalidated, and neither will any random sequence
+        that was forked off from it.
+        """
+        result = getattr(self, "_random_sequence_root", None)
+        if result is None:
+            self._random_sequence_root = RandomSequence(seed=self.random_seed)
+        return self._random_sequence_root