+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Running instances of Skybrush Viewer are now detected correctly on macOS even
+  if the user is not connected to any network at all. (The issue may have
+  affected Windows and/or Linux as well, but we encountered it on macOS).
+
 ## [2.4.0] - 2022-08-21
 
 ### Added